Level 7 updated
In this version, the concept of more OOP is implemented. Where the various classes are separated from the main Chatbot. Chatbot is also removed.
