+++ src/main/java/UI.java
@@ -1,0 +1,2 @@
+public class UI {
+}
